Create new grid_map_utils package and refactor pure pursuit
