+++ urc_navigation/grid_map_utils/include/grid_map_utils/grid_map_utils.hpp
@@ -1,0 +1,29 @@
+#ifndef GRID_MAP_UTILS__GRID_MAP_UTILS_HPP_
+#define GRID_MAP_UTILS__GRID_MAP_UTILS_HPP_
+
+#include <string>
+
+#include <grid_map_msgs/msg/grid_map.hpp>
+
+namespace grid_map_utils
+{
+class GridMapUtils
+{
+public:
+  GridMapUtils();
+
+  void setMap(const grid_map_msgs::msg::GridMap & grid_map);
+  void setLayer(const std::string & layer_name);
+
+  int getLayerIndex() const;
+  bool getMapDimensions(int & width, int & height) const;
+  bool worldToGrid(double x, double y, int & map_x, int & map_y) const;
+  bool tryGetCellCost(double x, double y, float & cost) const;
+
+private:
+  grid_map_msgs::msg::GridMap grid_map_;
+  std::string layer_name_;
+};
+}  // namespace grid_map_utils
+
+#endif  // GRID_MAP_UTILS__GRID_MAP_UTILS_HPP_
