+++ urc_navigation/path_planning/include/astar.hpp
@@ -6,7 +6,7 @@
 #include <urc_msgs/msg/grid_location.hpp>
 #include <geometry_msgs/msg/pose.hpp>
 #include <grid_map_msgs/msg/grid_map.hpp>
-#include "grid_map_utils.hpp"
+#include "grid_map_utils/grid_map_utils.hpp"
 #include <math.h>
 #include <unistd.h>
 #include <queue>
