+++ urc_navigation/trajectory_following/config/pure_pursuit_config.yaml
@@ -6,7 +6,7 @@ follower_action_server:
     heading_alignment_tolerance: 0.2
     enable_holonomic_motion: true
     lethal_cost_threshold: 0.5
-    cmd_vel_stamped: true
+    cmd_vel_stamped: false
     cmd_vel_topic: "/cmd_vel"
     map_frame: "map"
     base_link_frame: "base_link"
